GetOrCreateAccessionWrapperCreator made generic + moved to core module
